Start migrating evaluation classes
